requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
anyio==4.4.0
    # via
    #   httpx
    #   jupyter-server
appnope==0.1.4
    # via ipykernel
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.16.0
    # via jupyterlab-server
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.1.0
    # via nbconvert
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.0
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.5
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
exceptiongroup==1.2.2
    # via
    #   anyio
    #   ipython
executing==2.0.1
    # via stack-data
fastjsonschema==2.20.0
    # via nbformat
fonttools==4.53.1
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via jupyterlab
idna==3.7
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
importlib-metadata==8.4.0
    # via
    #   jupyter-client
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
importlib-resources==6.4.4
    # via matplotlib
ipykernel==6.29.5
    # via
    #   project-template (pyproject.toml)
    #   jupyter
    #   jupyter-console
    #   jupyterlab
ipython==8.18.1
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==8.1.3
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.9.25
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema==4.23.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter==1.1.1
    # via project-template (pyproject.toml)
jupyter-client==8.6.2
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.10.0
    # via jupyter-server
jupyter-lsp==2.2.5
    # via jupyterlab
jupyter-server==2.14.2
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.2.4
    # via
    #   jupyter
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.11
    # via ipywidgets
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.9.2
    # via
    #   project-template (pyproject.toml)
    #   seaborn
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mistune==3.0.2
    # via nbconvert
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.4
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.2.1
    # via jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==2.0.2
    # via
    #   project-template (pyproject.toml)
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
overrides==7.7.0
    # via jupyter-server
packaging==24.1
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
pandas==2.2.3
    # via
    #   project-template (pyproject.toml)
    #   seaborn
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.4.0
    # via matplotlib
platformdirs==4.2.2
    # via jupyter-core
prometheus-client==0.20.0
    # via jupyter-server
prompt-toolkit==3.0.47
    # via
    #   ipython
    #   jupyter-console
psutil==6.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-json-logger==2.0.7
    # via jupyter-events
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via jupyter-events
pyzmq==26.2.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.3
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
seaborn==0.13.2
    # via project-template (pyproject.toml)
send2trash==1.8.3
    # via jupyter-server
setuptools==73.0.1
    # via jupyterlab
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.6
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.3.0
    # via nbconvert
tomli==2.0.1
    # via jupyterlab
tornado==6.4.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20240821
    # via arrow
typing-extensions==4.12.2
    # via
    #   anyio
    #   async-lru
    #   ipython
tzdata==2024.1
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.2.2
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==24.8.0
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
widgetsnbextension==4.0.11
    # via ipywidgets
zipp==3.20.0
    # via
    #   importlib-metadata
    #   importlib-resources

flask==2.0.1
flask-restx==1.3.0
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
gunicorn==20.1.0